+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `new_unchecked` method which skips string checking.
+
 ## [0.1.1] - 2021-07-21
 ### Added
 - Convenient conversion to uppercase and lowercase hexadecimal string.