+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- Bump Rust edition to 2021
+
+## [0.1.2] - 2021-08-11
+### Added
+- `new_unchecked` method which skips string checking.
+
+### Fixed
+- Update CHANGELOG links
+
 ## [0.1.1] - 2021-07-21
 ### Added
 - Convenient conversion to uppercase and lowercase hexadecimal string.
@@ -16,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Convenient type aliases `UpperHexString` and `LowerHexString`.
 - Feature flag `serde` for serde support on `HexString` type.
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.1...HEAD
-[0.1.1]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.1.1
-[0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.1.0
+[Unreleased]: https://github.com/alekece/hextring-rs/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/alekece/hexstring-rs/releases/tag/v0.1.2
+[0.1.1]: https://github.com/alekece/hexstring-rs/releases/tag/v0.1.1
+[0.1.0]: https://github.com/alekece/hexstring-rs/releases/tag/v0.1.0